requirements.txt
fastapi
uvicorn
youtube_dl
supabase
python-dotenv
yt_dlp